+++ packages/author-profile/author-profile.stories.js
@@ -7,7 +7,7 @@ import { MockedProvider } from "@times-components/utils/graphql";
 import { addTypenameToDocument } from "apollo-utilities";
 import { query as authorProfileQuery } from "@times-components/provider/author-profile-provider";
 import { query as articleListQuery } from "@times-components/provider/article-list-provider";
-import storybookReporter from "@times-components/tealium";
+import storybookReporter from "@times-components/tealium/storybook";
 import AuthorProfile from "./author-profile";
 import AuthorProfileContent from "./author-profile-content";
 import authorProfileFixture from "./fixtures/author-profile.json";
